requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.112.2
fastapi-cli==0.0.5
Flask==3.0.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.7
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.1
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
redis==5.0.8
rich==13.8.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.4
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
Werkzeug==3.0.4
